requirements.txt
Faker==24.11.0
fastapi==0.110.2
langchain==0.1.16
langchain-experimental==0.0.57
langchain-openai==0.1.3
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
python-dotenv==1.0.1
spacy==3.7.4
uvicorn==0.29.0